+++ frontend/public/components/secrets/create-secret/const.ts
@@ -1,18 +1,1 @@
-export enum SecretTypeAbstraction {
-  generic = 'generic',
-  source = 'source',
-  image = 'image',
-  webhook = 'webhook',
-}
-
 export const AUTHS_KEY = 'auths';
-
-export enum SecretType {
-  basicAuth = 'kubernetes.io/basic-auth',
-  dockercfg = 'kubernetes.io/dockercfg',
-  dockerconfigjson = 'kubernetes.io/dockerconfigjson',
-  opaque = 'Opaque',
-  serviceAccountToken = 'kubernetes.io/service-account-token',
-  sshAuth = 'kubernetes.io/ssh-auth',
-  tls = 'kubernetes.io/tls',
-}
