+++ frontend/public/components/secrets/create-secret/index.tsx
@@ -1,17 +1,19 @@
-export * from './utils';
 export * from './const';
-export * from './SecretFormWrapper';
-export * from './PullSecretForm';
-export * from './PullSecretCredentialEntry';
-export * from './PullSecretCredentialsForm';
-export * from './PullSecretUploadForm';
-export * from './WebHookSecretForm';
-export * from './SourceSecretForm';
+export * from './types';
+export * from './utils';
+
 export * from './BasicAuthSubform';
-export * from './SSHAuthSubform';
-export * from './GenericSecretForm';
-export * from './KeyValueEntryForm';
-export * from './EditSecret';
 export * from './CreateSecret';
 export * from './DropableFileInput';
+export * from './EditSecret';
+export * from './GenericSecretForm';
+export * from './KeyValueEntryForm';
+export * from './PullSecretCredentialEntry';
+export * from './PullSecretCredentialsForm';
+export * from './PullSecretForm';
+export * from './PullSecretUploadForm';
+export * from './SecretFormWrapper';
 export * from './SecretSubForm';
+export * from './SourceSecretForm';
+export * from './SSHAuthSubform';
+export * from './WebHookSecretForm';
